+++ CHANGELOG.md
@@ -36,10 +36,23 @@
 * Use new API for DataImport landing page (UIDATIMP-918)
 * Enable skipped test for <ExistingSection> component and check overall test coverage of ui-data-import (UIDATIMP-1065)
 * Use new API for DataImport ViewAll page (UIDATIMP-1069)
+* Field mapping profiles created by auto tests couldn't be viewed (UIDATIMP-1073)
+* Remove 'react-dropzone' dependency (UIDATIMP-1072)
+* Field mapping profiles created by auto tests couldn't be viewed (UIDATIMP-1073)
+* Enable Action profile action for incoming MARC Authorities: Update MARC Authority records (UIDATIMP-1055)
+* Enable Matching profile for incoming MARC Authorities (UIDATIMP-1054)
+* Enable Field Mapping profile action for updating MARC Authorities records (UIDATIMP-1056)
+* Job profiles are showing incorrect info for "Jobs using this profile" accordion (UIDATIMP-1081)
+* Settings > Data Import > change focus (UIDATIMP-1085)
+* When duplicating job profile, retain the details from the profile being duplicated (UIDATIMP-1082)
 
 ### Bugs fixed:
 * Need to be able to use the same match profile in the same job profile, with different actions (UIDATIMP-1067)
 * The log hotlinks for Instance, Holdings, and/or Item record are not populated (UIDATIMP-1101)
+
+## [5.0.4](https://github.com/folio-org/ui-data-import/tree/v5.0.4) (2022-02-09)
+### Bugs fixed:
+* Need to be able to nest match profiles under action profiles (Kiwi HF) (UIDATIMP-1086)
 
 ## [5.0.3](https://github.com/folio-org/ui-data-import/tree/v5.0.3) (2021-11-25)